Implemented a controller daemon which reads /etc/linux_diagnostics/config.yaml. Depending on how the config is setup and which filesystems are mounted (cifs/nfs), the corresponding anomaly monitors are executed. Each of these monitors are polled at a granularity of 1 sec. If either of them trips, the diagnostics command is called and a tarball is generated with the report.
The controller also rotates the tarballs that are kept depending upon the number
of tarballs and disk usage.

Optionally, cpu and memory limits can be configured. The controller daemon will
put itself into a cgroup and restrict the cpu and memory usage.

Also added spec files for rpm and deb packages.

Signed-off-by: Shyam Prasad N <sprasad@microsoft.com>